+++ react/components/PageViewPixel.js
@@ -1,41 +1,39 @@
-import PropTypes from 'prop-types'
-import { Component } from 'react'
+import { useRef } from 'react'
+import { usePixel } from 'vtex.pixel-manager/PixelContext'
+import { useRuntime } from 'vtex.render-runtime'
 
-import { withRuntimeContext } from 'vtex.render-runtime'
-import { Pixel } from 'vtex.pixel-manager/PixelContext'
+// Page view events for home, search and product pages are sent by their own wrappers!
+const BLACKLIST_PAGES = ['store.home', 'store.search', 'store.product']
 
-class PageViewPixel extends Component {
-  static propTypes = {
-    push: PropTypes.func.isRequired,
-    runtime: PropTypes.object.isRequired,
-  }
+/** When you navigate to a new page, the route object will remain the one from the previous route.
+ * So we save the ref on the first mount. After that, we wait for the first time route changes to send a page view with the correct title.
+ * A possible better implementation for the future, is add a loading state in the route object.
+ */
 
-  sendPageViewEvent() {
-    this.props.push({
+const PageViewPixel = ({ title }) => {
+  const { push } = usePixel()
+  const { route, page, account } = useRuntime()
+
+  const routeRef = useRef(route)
+  const isBlacklisted =
+    BLACKLIST_PAGES.findIndex(elem => page.includes(elem)) >= 0
+
+  if (routeRef.current !== route && !isBlacklisted) {
+    routeRef.current = route
+    const data = {
       event: 'pageView',
-      pageTitle: document.title,
+      pageTitle: title,
       pageUrl: location.href,
       referrer:
         document.referrer.indexOf(location.origin) === 0
           ? undefined
           : document.referrer,
-      accountName: this.props.runtime.account,
-    })
-  }
-
-  componentDidMount() {
-    this.sendPageViewEvent()
-  }
-
-  componentDidUpdate(prevProps) {
-    if (prevProps.runtime.route.path !== this.props.runtime.route.path) {
-      this.sendPageViewEvent()
+      accountName: account,
     }
+    push(data)
   }
 
-  render() {
-    return null
-  }
+  return null
 }
 
-export default Pixel(withRuntimeContext(PageViewPixel))
+export default PageViewPixel